% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diagnostic_plots.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Creates plots that are useful when testing AgroIBS.}
\usage{
create_diagnostic_plots(model_path, create_interactive = F, save_dir = NULL)
}
\arguments{
\item{model_path}{Path to the model directory}

\item{create_interactive}{Logical, indicating if you want to create interactive plots which will be output has html files}

\item{save_dir}{If create_interactive is TRUE, this is the directory where the interactive plots will be saved.}
}
\value{
A ggplot object
}
\description{
Creates plots that are useful when testing AgroIBS.
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
