% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ibis_ts_plot.R
\name{create_ibis_ts_plot}
\alias{create_ibis_ts_plot}
\title{Creates a general timeseries plot from AgroIBIS output.}
\usage{
create_ibis_ts_plot(data_tbl)
}
\arguments{
\item{data_tbl}{Output from get_*_output functions}
}
\value{
A ggplot object
}
\description{
Creates a general timeseries plot from AgroIBIS output.
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
