% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weather_plot.R
\name{create_weather_plot}
\alias{create_weather_plot}
\title{Creates a plot that shows daily max and min temperature and precipitation}
\usage{
create_weather_plot(model_path)
}
\arguments{
\item{model_path}{Path to the model directory}
}
\value{
A ggplot object
}
\description{
Creates a plot that shows daily max and min temperature and precipitation
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
