% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_output.R
\name{get_yearly_output}
\alias{get_yearly_output}
\title{Gets yearly output and puts it into a tibble for analysis or plotting}
\usage{
get_yearly_output(filepath, ncvar, pft = NULL, average_spatial = T)
}
\arguments{
\item{filepath}{Path to the daily netCDF file}

\item{ncvar}{Name of variable of interest}

\item{pft}{Select the pft of interest. The default is set to NULL}

\item{average_spatial}{If set equal to false, the variable is not averaged over space}
}
\description{
Gets yearly output and puts it into a tibble for analysis or plotting
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
