% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ibis.R
\name{run_ibis}
\alias{run_ibis}
\title{Run Agro-IBIS and copy model output and restart files}
\usage{
run_ibis(
  model_path = ".",
  compile_flag = F,
  copy_output_path = NULL,
  copy_restart_path = NULL
)
}
\arguments{
\item{model_path}{Path to the model directory.}

\item{compile_flag}{If set to TRUE, it will compile the model before running.}

\item{copy_output_path}{Path to copy the output directory.}

\item{copy_restart_path}{Path to copy the restart directory.}
}
\description{
Run Agro-IBIS and copy model output and restart files
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
