% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_management.R
\name{write_management}
\alias{write_management}
\title{Writes planting date data to NetCDF file}
\usage{
write_management(model_path, yr, doy)
}
\arguments{
\item{model_path}{Path to the model directory.}

\item{yr}{Year or vector of years to change planting date.}

\item{doy}{Day of year to change planting date for the specific year(s).}
}
\description{
Writes planting date data to NetCDF file
}
\author{
Bryan Petersen - bryan20@iastate.edu
}
