% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ibis.R
\name{setup_ibis}
\alias{setup_ibis}
\title{Set the ibis.infile}
\usage{
setup_ibis(
  model_path,
  file = "ibis_flag.infile",
  irestart = 1,
  irstyear,
  iyear0 = 1751,
  nrun,
  soilcspin = 0,
  flg_wrestart = 1,
  iyearout = 1,
  imonthout = 0,
  idailyout = 0,
  ihourlyout = 0,
  isimveg,
  nstress = 1,
  imiscanthus0,
  irotation,
  iholdsoiln = 1,
  co2init,
  snorth,
  ssouth,
  swest,
  seast
)
}
\arguments{
\item{model_path}{Path to the model directory.}

\item{file}{Name of the ibis_flag.infile}

\item{irestart}{0: not a restart run  1: restart run}

\item{irstyear}{actual calendar year of restart run}

\item{iyear0}{initial year of simulation (don't change for restart)}

\item{nrun}{number of years in this simulation (change for restart)}

\item{soilcspin}{0: no soil spinup, 1: acceleration procedure used}

\item{flg_wrestart}{1(default): write restart files  0: no restart files}

\item{iyearout}{0: no yearly output, 1: yearly output   2: yearly output,  in a yearly basis}

\item{imonthout}{0: no monthly output, 1: monthly output 2: monthly output, in a yearly basis}

\item{idailyout}{0: no daily output, 1: daily output}

\item{ihourlyout}{0: no hourly output, 1: hourly output}

\item{isimveg}{0: static veg, 1: dynamic veg, 2: dynamic veg-cold start}

\item{nstress}{0: no nitrogen stress, 1: apply nitrogen stress}

\item{irotation}{0: none -- 1: w. wheat/fallow -- 2: corn/soy -- 3: corn/corn/soy -- 4: soy/w. wheat/corn -- 5: soy/corn (opposite of 2) -- 6: biomass sorghum/soybean}

\item{iholdsoiln}{0: doesn't save soil inorganic N from restart 1: save inorganic soil N}

\item{co2init}{initial co2 concentration in mol/mol (ex: 0.000400)}

\item{snorth}{northern latitude for subsetting in/output}

\item{ssouth}{southern latitude for subsetting in/output}

\item{swest}{western longitude for subsetting in/output}

\item{seast}{eastern longitude for subsetting in/output}

\item{imiscanthus}{0: miscanthus not grown 1: miscanthus grown everywhere}
}
\value{
ibis.infile
}
\description{
Set the ibis.infile
}
